--- src/app/chat/style.ts
@@ -1,10 +1,0 @@
-import { createStyles } from 'antd-style';
-
-export const useStyles = createStyles(({ css }) => ({
-  header: css`
-    position: absolute;
-    z-index: 1;
-    top: 0;
-    left: 0;
-  `,
-}));
