+++ src/app/market/Market/index.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { createStyles } from 'antd-style';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
