+++ src/app/market/layout.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { ReactNode, memo } from 'react';
 import { Flexbox } from 'react-layout-kit';
