+++ src/components/GridList/index.tsx
@@ -1,10 +1,6 @@
-import { Icon } from '@lobehub/ui';
-import { Empty, Space } from 'antd';
+import { Skeleton } from 'antd';
 import classNames from 'classnames';
-import { Loader2 } from 'lucide-react';
 import React from 'react';
-import { useTranslation } from 'react-i18next';
-import { Center } from 'react-layout-kit';
 
 import ListItem from './ListItem';
 import { useStyles } from './style';
@@ -17,9 +13,6 @@ interface Item {
 
 interface GridListProps {
   className?: string;
-  empty?: {
-    actions?: React.ReactNode[];
-  };
   isActivated?: (id: string) => boolean;
   isChecked?: (id: string) => boolean;
   items: any[];
@@ -28,29 +21,23 @@ interface GridListProps {
   style?: React.CSSProperties;
 }
 
-const GridList = (props: GridListProps) => {
-  const {
-    items,
-    className,
-    style,
-    onClick,
-    isActivated,
-    isChecked,
-    loading = false,
-    empty,
-  } = props;
-  const { styles } = useStyles();
-  const { t } = useTranslation('common');
-
-  const Loading = () => (
-    <Center gap={16} horizontal className={styles.loading}>
-      <Icon icon={Loader2} spin />
-      {t('loading')}
-    </Center>
+const SkeletonList = () => {
+  const total = 18;
+  return (
+    <>
+      {Array.from({ length: total }).map((_, i) => (
+        <Skeleton.Image active={true} key={i} />
+      ))}
+    </>
   );
+};
+
+const GridList = (props: GridListProps) => {
+  const { items, className, style, onClick, isActivated, isChecked, loading = false } = props;
+  const { styles } = useStyles();
 
   const List = () => (
-    <div className={styles.list}>
+    <>
       {items.map((item) => {
         const { avatar, name, id } = item;
         return (
@@ -66,18 +53,12 @@ const GridList = (props: GridListProps) => {
           />
         );
       })}
-    </div>
-  );
-
-  const EmptyList = () => (
-    <Empty image={Empty.PRESENTED_IMAGE_SIMPLE} description={t('noData')}>
-      {empty?.actions ? <Space>{empty.actions}</Space> : null}
-    </Empty>
+    </>
   );
 
   return (
     <div className={classNames(className, styles.grid)} style={style}>
-      {loading ? <Loading /> : items.length === 0 ? <EmptyList /> : <List />}
+      <div className={styles.list}>{loading ? <SkeletonList /> : <List />}</div>
     </div>
   );
 };
